fix(extract): parse mph/none/knots maxspeed — 96-100% of US/GB majors had no speed
parse::<u8>() silently dropped every non-bare-integer maxspeed: '70 mph'
(whole US/GB tagging — taginfo: '30 mph' alone is 3.96% of all maxspeed
values), 'none' (52.8k derestricted Autobahn ways), knots, and any value
>255 (rail '125 mph' became 0 -> highspeed fallback 300 km/h; audit R3,
Tring WCML +5.3 dB).

- osm-extract: parse_maxspeed_kmh (mph x1.609, knots x1.852, walk->10,
  none->MAXSPEED_NONE=u16::MAX, signals/garbage->0, clamp <=400); road
  column stays u8 (none->255 SPEED_LIMIT_DERESTRICTED, real values
  min(254) — collision-free); rail column u8->u16 (300+ km/h survives)
- tolerant readers (old all-u8 arrows stay readable — they ARE the
  migration): source-reader col_u16_or_u8, heatmap-aircraft u16-else-u8
- normalize: 255 -> DERESTRICTED_SPEED_KMH = 130 (BASt 2025 measured
  mean 124.1 km/h on derestricted BAB; CNOSSOS validity cap)
- metadata/UI (review W4): speed_posted_kmh u8 -> Option<u8>, null =
  derestricted, popup renders 'no limit' (keyed off null — junction-
  independent); RailMetadata u16
- SPEC notes (incl. the C8 dwell-sign phrase, committed together here)

Zero popup drift: scanned all 464,735,138 local road rows — none carry
255; legacy 'none'/'mph' already degraded to 0 (class default) and still
do until the next OSM re-extract heals the data. he84 mirror step:
sampled 255-scan before restamp. /ritual: simplify + Codex/Gemini, 0
CRITICAL; Codex's dataset-scope caveat recorded here.

+++ engine/noise-compute/src/types/inputs.rs
@@ -60,7 +60,7 @@ pub struct RoadSegment {
     pub end_lon: f64,
     pub length_m: f32,
     pub road_class: u8,   // 0=motorway..6=living_street, 7=service, 8=track, 9=unclassified, 10=motorway_link, 11=trunk_link, 12=primary_link
-    pub speed_limit: u8,  // km/h, 0=use default
+    pub speed_limit: u8,  // km/h, 0=use default, 255=derestricted (maxspeed=none) → DERESTRICTED_SPEED_KMH
     pub surface_type: u8, // 0=asphalt..4=gravel
     pub oneway: bool,
     pub lanes: u8,
@@ -94,7 +94,7 @@ pub struct RailSegment {
     pub length_m: f32,
     pub rail_type: u8, // 0=rail, 1=tram, 2=light_rail, 3=narrow_gauge, 4=funicular
     pub usage: u8,     // 0=main, 1=branch, 2=industrial
-    pub maxspeed: u8,  // km/h (raw OSM value, 0 = none)
+    pub maxspeed: u16, // km/h (raw OSM value, 0 = none); u16 so 300+ km/h survives
     pub trains_passenger: f64, // effective daily count (post service/divisor scaling)
     pub trains_freight: f64, // effective daily count (post service/divisor scaling)
     pub speed_kmh: f64, // effective speed used by emission (resolved); f64 not u8 — high-speed rail resolves to 300 km/h, which u8 saturated to 255 (~1.4 dB too quiet)
